@@ -13,15 +13,15 @@
         }
     ],
      "require": {
-        "php": "~7.1",
+        "php": "~7.4",
         "silex/silex": "~2.0",
         "guzzle/guzzle": ">=3.8",
         "symfony/process": "~3|~4|~5",
         "jeremeamia/superclosure": "~2",
-        "lstrojny/hmmmath": ">=0.5.0"
+        "lstrojny/hmmmath": ">=0.5.0",
+        "slevomat/coding-standard": "^7.0"
     },
     "require-dev": {
-        "internations/kodierungsregelwerksammlung": "~0.23.0",
         "internations/testing-component": "1.0.1",
         "phpunit/phpunit": "^7"
     },
