@@ -5,14 +5,14 @@ use Closure;
 
 class ClosureMatcher extends AbstractMatcher
 {
-    private $closure;
+    private Closure $closure;
 
     public function __construct(Closure $closure)
     {
         $this->closure = $closure;
     }
 
-    protected function createMatcher()
+    protected function createMatcher(): Closure
     {
         return $this->closure;
     }
