@@ -15,7 +15,7 @@ class RequestStorage
         $this->directory = $directory;
     }
 
-    public function store(Request $request, $name, $data)
+    public function store(Request $request, $name, $data): void
     {
         file_put_contents($this->getFileName($request, $name), serialize($data));
     }
@@ -31,14 +31,14 @@ class RequestStorage
         return Util::deserialize(file_get_contents($fileName));
     }
 
-    public function append(Request $request, $name, $data)
+    public function append(Request $request, $name, $data): void
     {
         $list = $this->read($request, $name);
         $list[] = $data;
         $this->store($request, $name, $list);
     }
 
-    public function prepend(Request $request, $name, $data)
+    public function prepend(Request $request, $name, $data): void
     {
         $list = $this->read($request, $name);
         array_unshift($list, $data);
@@ -50,7 +50,7 @@ class RequestStorage
         return $this->directory . $this->pid . '-' . $name . '-' . $request->server->get('SERVER_PORT');
     }
 
-    public function clear(Request $request, $name)
+    public function clear(Request $request, $name): void
     {
         $fileName = $this->getFileName($request, $name);
 
