@@ -7,12 +7,12 @@ class CallbackResponse extends Response
 {
     private $callback;
 
-    public function setCallback(callable $callback)
+    public function setCallback(callable $callback): void
     {
         $this->callback = $callback;
     }
 
-    public function sendCallback()
+    public function sendCallback(): void
     {
         if ($this->callback) {
             $callback = $this->callback;
@@ -20,7 +20,7 @@ class CallbackResponse extends Response
         }
     }
 
-    public function send()
+    public function send(): void
     {
         $this->sendCallback();
         parent::send();
